LightningQr 컴포넌트를 정의한다
